fix: Removed compiler warnings about implicit casts (Xcode 12 iOS)

+++ interface/basisu_c_binding/inc/basisu_c_binding.h
@@ -44,7 +44,7 @@ public:
 extern "C" {
 DLL_EXPORT void ktx_basisu_basis_init();
 DLL_EXPORT basis_file* ktx_basisu_create_basis();
-DLL_EXPORT bool ktx_basisu_open_basis( basis_file* basis, const uint8_t * data, size_t length );
+DLL_EXPORT bool ktx_basisu_open_basis( basis_file* basis, const uint8_t * data, uint32_t length );
 DLL_EXPORT void ktx_basisu_close_basis( basis_file* basis );
 DLL_EXPORT void ktx_basisu_delete_basis( basis_file* basis );
 DLL_EXPORT bool ktx_basisu_getHasAlpha( basis_file* basis );
@@ -54,6 +54,6 @@ DLL_EXPORT uint32_t ktx_basisu_getImageWidth( basis_file* basis, uint32_t image_
 DLL_EXPORT uint32_t ktx_basisu_getImageHeight( basis_file* basis, uint32_t image_index, uint32_t level_index);
 DLL_EXPORT uint32_t ktx_basisu_getImageTranscodedSizeInBytes( basis_file* basis, uint32_t image_index, uint32_t level_index, uint32_t format);
 DLL_EXPORT bool ktx_basisu_startTranscoding( basis_file* basis );
-DLL_EXPORT bool ktx_basisu_transcodeImage( basis_file* basis, void* dst, size_t dst_size, uint32_t image_index, uint32_t level_index, uint32_t format, uint32_t pvrtc_wrap_addressing, uint32_t get_alpha_for_opaque_formats);
+DLL_EXPORT bool ktx_basisu_transcodeImage( basis_file* basis, void* dst, uint32_t dst_size, uint32_t image_index, uint32_t level_index, uint32_t format, uint32_t pvrtc_wrap_addressing, uint32_t get_alpha_for_opaque_formats);
 }
 #endif
